+++ src/main/java/com/powsybl/openloadflow/network/PiModel.java
@@ -55,27 +55,6 @@ public interface PiModel {
 
     void roundR1ToClosestTap();
 
-    enum Direction {
-        INCREASE(AllowedDirection.INCREASE),
-        DECREASE(AllowedDirection.DECREASE);
-
-        private final AllowedDirection allowedDirection;
-
-        Direction(AllowedDirection allowedDirection) {
-            this.allowedDirection = allowedDirection;
-        }
-
-        public AllowedDirection getAllowedDirection() {
-            return allowedDirection;
-        }
-    }
-
-    enum AllowedDirection {
-        INCREASE,
-        DECREASE,
-        BOTH
-    }
-
     boolean updateTapPositionA1(Direction direction);
 
     Optional<Direction> updateTapPositionR1(double deltaR1, int maxTapShift, AllowedDirection allowedDirection);
